pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "growthlint"
version = "0.1.0"
description = "The ESLint for growth marketing. Scan repos and websites for conversion leaks, missing analytics, broken attribution, and SEO issues."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "GrowthLint Contributors" }]
keywords = ["growth", "marketing", "conversion", "analytics", "seo", "linting", "audit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "lxml>=5.1.0",
    "defusedxml>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
growthlint = "growthlint.cli:app"

[project.urls]
Homepage = "https://github.com/super0510/GrowthLint"
Repository = "https://github.com/super0510/GrowthLint"
Issues = "https://github.com/super0510/GrowthLint/issues"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
